+++ src/IO.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 namespace Innmind\IO;
 
 use Innmind\IO\Internal\Capabilities;
-use Innmind\TimeContinuum\Clock;
+use Innmind\Time\Clock;
 
 final class IO
 {