+++ src/Internal/Capabilities/Async.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 namespace Innmind\IO\Internal\Capabilities;
 
 use Innmind\IO\Internal\Watch;
-use Innmind\TimeContinuum\Clock;
+use Innmind\Time\Clock;
 
 /**
  * @internal