+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Requires PHP `8.4`
 - Requires `innmind/immutable:~6.0`
+- Requires `innmind/time:~1.0`
 
 ## 3.5.1 - 2025-08-18